+++ core/src/registration_plane_feature.cpp
@@ -11,16 +11,16 @@
 #include <python-scripts/point-to-feature-metrics/point_to_plane_tait_bryan_wc_jacobian.h>
 #include <python-scripts/point-to-point-metrics/point_to_projection_onto_plane_tait_bryan_wc_jacobian.h>
 
-std::vector<RegistrationPlaneFeature::Job> RegistrationPlaneFeature::get_jobs(long long unsigned int size, int num_threads)
+std::vector<RegistrationPlaneFeature::Job> RegistrationPlaneFeature::get_jobs(uint64_t size, int num_threads)
 {
     int hc = size / num_threads;
     if (hc < 1)
         hc = 1;
 
     std::vector<Job> jobs;
-    for (long long unsigned int i = 0; i < size; i += hc)
+    for (uint64_t i = 0; i < size; i += hc)
     {
-        long long unsigned int sequence_length = hc;
+        uint64_t sequence_length = hc;
         if (i + hc >= size)
         {
             sequence_length = size - i;
